+++ changelog.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Perform new borrowing feature test.
     * Retrieve borrowing feature test.
     * "New borrowing" page unit tests.
+    * "End borrowing" page unit tests.
 * Validation
     * Added maximal value for borrowing guarantee.
 * End borrowing view